RAG/requirements.txt
# Core RAG dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

# Embeddings
sentence-transformers==2.7.0

# Vector Database
chromadb==0.4.22

# Text processing
tiktoken==0.5.2
pypdf==3.17.4

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# Web Interface
streamlit==1.28.0

# Optional: For better text splitting
unstructured==0.11.8

